+++ src/test/java/kitchenpos/application/fixture/ProductFixture.java
@@ -8,7 +8,7 @@ public class ProductFixture {
     private ProductFixture() {
     }
 
-    public static Product create(final Long id, final String name, final Long price) {
-        return new Product(id, name, BigDecimal.valueOf(price));
+    public static Product create(final Long id, final String name, final BigDecimal price) {
+        return new Product(id, name, price);
     }
 }
